delivery
delivery controller

+++ src/main/java/com/swd/evdms/repository/OrderRepository.java
@@ -2,9 +2,8 @@ package com.swd.evdms.repository;
 
 import com.swd.evdms.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
     List<Order> findByUser_UserId(Integer userId);}
